+++ src/main/java/project/token/IntToken.java
@@ -1,15 +1,17 @@
 package project.token;
 
+import java.math.BigInteger;
+
 public class IntToken extends Token {
 
-    private final int value;
+    private final BigInteger value;
 
-    public IntToken(TokenType type, int position, int lineNr, int positionAtLine, int value) {
+    public IntToken(TokenType type, int position, int lineNr, int positionAtLine, BigInteger value) {
         super(type, position, lineNr, positionAtLine);
         this.value = value;
     }
 
-    public int getValue() {
+    public BigInteger getValue() {
         return value;
     }
 }
